+++ tests/application/test_forms.py
@@ -5,7 +5,7 @@ from mock import Mock
 import os
 import json
 from application.dashboards import build_dict_for_post
-from tests.admin.test_dashboards import data_sets_list
+from tests.application.test_dashboards import data_sets_list
 
 
 class DashboardTestCase(TestCase):
